x87cmp: three float comparisons had the wrong operator
gcc 2.7.2.3's `fnstsw ax` epilogue encodes the comparison OPERATOR, so a
float `if` written the wrong way round is visible in the instruction stream
even when both sides branch to the same place.  x87cmp.py (added last commit)
screens for it: 168 functions carry x87 compares, 142 agree, 26 differ.

The table was re-measured with probe_snippet.sh at OUR flags rather than

  - textual operand order is INERT (`b >= a` is byte-identical to `a <= b`,
    `b > a` to `a < b`), so a mask difference is a genuinely different
    OPERATOR -- which is what makes the screen worth running -- and "swap the
    operands to change the shape" is not a lever;
  - `if (!(a < b))` IS the same code as `if (a >= b)` here, because our flag
    set has -ffast-math and theirs does not.

Three fixes, all one class: IDA wrote `if (v < 0) a; else b;` where the
original is `if (v > 0) b; else a;`.  The two differ only at exactly 0.0,
which is why nothing else ever caught them.

  BotRoamGoal (be_ai2_dmq2.c)   two `>= 0.5` -> `> 0.5`
                                OUR+8,  insn_diffs 192 -> 188
  sub_10025070 (be_ai2_dmq2.c)  the same loop, same fix

Also removed IDA's phantom `(float)` cast in Characteristic_BFloat's guard
(both operands are already float; measured a byte-exact no-op).

Measured on both oracles, FORCE=1 clean rebuild each:
  ELF botlib  MATCH 697 unchanged, total insn_diffs 25378 -> 25333,
              no row other than the three moved.
  PE          MATCH 757 unchanged, no status change anywhere; the three rows
              stay INSN_COUNT_MATCH and pay 5 bytes between them
              (Sum byte_diffs 51356 -> 51361).
Kept: the ELF is primary, PE instruction parity is untouched on all three
fidelity fixes read straight off the binary.

Two MASK hits investigated and deliberately NOT fixed:
stack as the error Print's %f args and emits `fcom st(1)`; ours reloads and
emits `fcomp <mem>` -- same operator, mirrored operands), and
AAS_ClientMovementPrediction compares against a computed value where ours
compares a .rodata constant, i.e. a different expression.  Both written up in
x87cmp.py's docstring so the next reader does not re-chase them.

PLAN.md: A4 resolved (F663 is the live botimport.Trace forwarder, F680 the
dead direct variant -- our #ifdef _WIN32 gives the Linux build the wrong
AAS_Trace body, which is a behavioural divergence for the box86 oracle too;
fix written up, not applied).  B2 triage recorded (15 of the 113 non-MATCH ELF
rows are content-complete, 25 have equal token counts, the rest differ in
AnglesToAxis pitch_rad experiment, is a re-runnable lever; the rest were note
staleness or scanner artifacts.  B3 left BLOCKED with the reason.

Source-commit: 072d7572e2250bb58d49ac0595e5800af839c35e

+++ botlib/be_ai_char.c
@@ -282,7 +282,7 @@ float __cdecl Characteristic_BFloat(bot_character_t *character, int index, float
   float result; // st7 — returns float (disasm loads fld DWORD / fcom DWORD,
                 // not the double-promotion sequence); Q3's is float too.
 
-  if ( min > (float)max )
+  if ( min > max )
   {
     botimport.Print(PRT_ERROR, "cannot bound characteristic %d between %f and %f\n", index, min, max);
     return 0.0f;
